requirements.txt
SpeechRecognition
openai-whisper
moviepy
pydub
PyAudio